refactor: 데이터 타입 및 사이즈 수정
- 데이터 크기를 최적화하여 디스크에 저장되는 데이터 양 절약
- int 대신 tinyint 사용
- text 대신 varchar 사용

+++ backend/was/src/tarot/entities/tarot-card.entity.ts
@@ -14,7 +14,7 @@ export class TarotCard {
   @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column('int')
+  @Column('tinyint')
   cardNo: number;
 
   @Column({ length: 10 })
